requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
neo4j>=5.15.0
ollama>=0.1.0
pytest>=7.0.0
requests>=2.28.0
PyYAML>=6.0
pdfplumber>=0.10.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
psutil>=5.9.0
spacy>=3.4.0
scikit-learn>=1.0.0
redis>=4.5.0
hiredis>=2.2.0
APScheduler>=3.10.0
docker
httpx
tavily-python